# Dashboard Performance: HtmlContent improvements