## What type of PR is this? (check all applicable)
<!-- Please leave only what's applicable -->

- [x] Other

## Description
Second PR for #4714, basically all widgets had this component rendered unnecessarily (also it could be memoized as its props should always return the same component). Although the `sanitize` function is kinda of performant, I believe it scales with the number of renders and widgets.

**Individual improvement (based on master)**
| Dashboard | master | This PR | % of improvement |
| --------------- | --------- | ---------- | ------------------------- |
| Sales Dashboard | 3599 ms | 2859 ms |  20% |
| Dashboard for testing React migration | 42505 ms | 37854 ms | 10% | 

_(The performance analysis were run with the same time for each dashboard twice (with the least value selected) to make sure results more reliable.)_

The Scripting time I'm noting actually varies considerably, but when there's an actual improvement, the minimum value you can get diminishes, so it's been the best benchmark I've found so far. It just makes it a bit more difficult to measure dashboards with low avg scripting time (< 3s).

## Related Tickets & Documents
#4714 

## Mobile & Desktop Screenshots/Recordings (if there are UI changes)
--